Fix spill drawing, use a fixed navbar
When drawing spills, we now use event.pageX and event.pageY as the X, Y
coordinates of the user click, since these values are relative to the
document. This fixes an issue with the existing code where event.clientX
and event.clientY would be different for the same position if the user
had scrolled the page.