added __eq__ for Timeseries class - keep this simple so it duck types and only compares output of   get_timeseries() - no longer setting 'name' and 'data_source' here - this is done by   derived class. Base class only manages + contains timeseries
Fixed bug where timeseries were not being compare in __eq__
- overloading base __eq__ so Wind object compares attributes being
  'saved' by invoking base __eq__
- it then also explicitly calls Timeseries.__eq__ to ensure the
  timeseries for two objects match
- moved the logic for initializing 'name' and 'source_data' here

added tests
currently tests in test_persist.py fail because the Wind data being
written out to timeseries has a mismatch in 4th decimal place (r,
theta). This is likely due to many conversions it goes through, but we
only need to keep and check 2 decimal places so update this.