added code so an OSSMTimeValue series will return very large and small numbers for a constant wind.
As close as I could think of for -InfTime and +InfTime

But it is now failing with an ou tof range error for making a time out of the value.

I used:

std::numeric_limits<Seconds>::max()

To get the max value, but I guess that isn't working (on 64 bit OS-X)

also -- I used 0 for minimum time, but maybe that should be negative some_very_large_value

Maybe this logic should be pushed to the Cyton/Python layer.

donpt merge this without resolving these issues!